Initial work on .deb packages
The basis of creating a .deb package is implemented with this commit. One can create a working .deb package by issuing ./make package deb. One hindering bug is that on installing the file attributes are lost. Need to fix this!
Also an undiscovered bug in the API procedure project was found. On cleaning up the version number not all of the string was searched.